Tile Object Editor Renaming
There is a tile object editor now that currently lets you rename a tile object, causing the file system dock to reload. Static typing is also now enforced and there is a new Constants class for global constants. Currently it only has DIRS, a list of the four orthogonal directions.